Move comms models out of model crate
