+++ rust/cloud-storage/comms_db_client/Cargo.toml
@@ -7,6 +7,7 @@ version = "0.1.0"
 [dependencies]
 anyhow = { workspace = true }
 cached = { workspace = true }
+channels = { path = "../channels", default-features = false, features = ["outbound"] }
 chrono = { workspace = true }
 doppleganger = { workspace = true }
 futures = { workspace = true }
