+++ rust/cloud-storage/seed_cli/Cargo.toml
@@ -7,6 +7,7 @@ version = "0.1.0"
 [dependencies]
 anyhow = { workspace = true }
 aws-sdk-s3 = { workspace = true }
+channels = { path = "../channels", default-features = false, features = ["outbound"] }
 chrono = { workspace = true }
 clap = { workspace = true }
 comms_db_client = { path = "../comms_db_client" }
